Move tag display functions to separate file
